// backend_new/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// @ts-nocheck 
/**
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/
export const UsersRole = {
  USER: 'USER',
  ADMIN: 'ADMIN',
  LIMITED_ADMIN: 'LIMITED_ADMIN',
  LIMITED_ADMIN_GMU: 'LIMITED_ADMIN_GMU'
} as const

export type UsersRole = (typeof UsersRole)[keyof typeof UsersRole]


export const LibraryType = {
  DOCUMENTATION: 'DOCUMENTATION',
  PRACTICE: 'PRACTICE'
} as const

export type LibraryType = (typeof LibraryType)[keyof typeof LibraryType]


export const InteractionType = {
  GMU: 'GMU',
  TNO: 'TNO'
} as const

export type InteractionType = (typeof InteractionType)[keyof typeof InteractionType]


export const ActionHistory = {
  INSERT: 'INSERT',
  UPDATE: 'UPDATE',
  DELETE: 'DELETE'
} as const

export type ActionHistory = (typeof ActionHistory)[keyof typeof ActionHistory]


export const ActivesType = {
  TRANSPORT: 'TRANSPORT',
  PROPERTY: 'PROPERTY',
  GROUND: 'GROUND',
  DEBIT: 'DEBIT',
  OTHER: 'OTHER'
} as const

export type ActivesType = (typeof ActivesType)[keyof typeof ActivesType]


export const DataStatus = {
  AIS: 'AIS',
  GMU: 'GMU',
  AIS_GMU: 'AIS_GMU',
  OLD_DATA: 'OLD_DATA'
} as const

export type DataStatus = (typeof DataStatus)[keyof typeof DataStatus]


export const ObjectStatus = {
  WANTED_BEGIN: 'WANTED_BEGIN',
  WANTED_END: 'WANTED_END',
  ARREST_BEGIN: 'ARREST_BEGIN',
  ARREST_END: 'ARREST_END',
  EVALUATION_BEGIN: 'EVALUATION_BEGIN',
  EVALUATION_END: 'EVALUATION_END',
  REALIZATION_FIRST_SUBMIT: 'REALIZATION_FIRST_SUBMIT',
  REALIZATION_FIRST: 'REALIZATION_FIRST',
  NOT_REALIZATION_FIRST: 'NOT_REALIZATION_FIRST',
  REALIZATION_SECOND_SUBMIT: 'REALIZATION_SECOND_SUBMIT',
  REALIZATION_SECOND: 'REALIZATION_SECOND',
  NOT_REALIZATION_SECOND: 'NOT_REALIZATION_SECOND',
  REFUND_PROPERTY: 'REFUND_PROPERTY',
  COMPLAINT: 'COMPLAINT',
  COMPLAINT_REVIEWED: 'COMPLAINT_REVIEWED',
  LEASING: 'LEASING',
  FORECLOSURE: 'FORECLOSURE',
  FORECLOSURE_CANCEL: 'FORECLOSURE_CANCEL',
  OTHER: 'OTHER'
} as const

export type ObjectStatus = (typeof ObjectStatus)[keyof typeof ObjectStatus]


export const LeasStatus = {
  IS_PLEDGE_HOLDER: 'IS_PLEDGE_HOLDER',
  IS_NOT_PLEDGE_HOLDER: 'IS_NOT_PLEDGE_HOLDER',
  NO_PLEDGE: 'NO_PLEDGE'
} as const

export type LeasStatus = (typeof LeasStatus)[keyof typeof LeasStatus]


export const WantedResults = {
  FINDING_PROPERTY: 'FINDING_PROPERTY',
  END_PROPERTY_SEARCH_ACTIVITIES: 'END_PROPERTY_SEARCH_ACTIVITIES'
} as const

export type WantedResults = (typeof WantedResults)[keyof typeof WantedResults]


export const RealizationStage = {
  FIRST: 'FIRST',
  SECOND: 'SECOND'
} as const

export type RealizationStage = (typeof RealizationStage)[keyof typeof RealizationStage]